requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly
geopandas
folium

# Statistical analysis
scipy>=1.10.0
statsmodels
prophet

# Machine learning
scikit-learn>=1.3.0
xgboost

# API and data fetching
requests>=2.31.0
wbgapi>=1.0.12
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
jupyterlab
ipykernel>=6.25.0

# Web apps
streamlit

# Optional but recommended
pyarrow>=12.0.0  # For Parquet file support
